refactor: simplify build and packaging of dropdown component
