+++ packages/components/dropdown/css/bs-dropdown-item-text.css.js
